requirements.txt
streamlit
google-generativeai
requests
pyTelegramBotAPI
scikit-learn
pandas